+++ src/knot_cloud.h
@@ -64,7 +64,7 @@ int knot_cloud_unregister_device(const char *id);
 int knot_cloud_auth_device(const char *id, const char *token);
 int knot_cloud_update_config(const char *id, struct l_queue *config_list);
 int knot_cloud_list_devices(void);
-int knot_cloud_publish_data(const char *id, uint8_t sensor_id,
+int knot_cloud_publish_data(const char *id, int sensor_id,
 			    uint8_t value_type, const knot_value_type *value,
 			    uint8_t kval_len, struct tm *ptm);
 int knot_cloud_read_start(const char *id, knot_cloud_cb_t read_handler_cb,
